Add boot function entrypoint
